# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Espace Show API Documentation",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/documentation",
            "description": "Espace Show API Server"
        }
    ],
    "paths": {
...
        "/api/allCountryAndPosition": {
            "get": {
                "tags": [
                    "PaysEtatsVilles"
                ],
                "summary": "Obtenir toutes les informations sur les pays et leurs positions",
                "operationId": "e05c563c1b12244e702f10df0654bc28",
                "responses": {
                    "200": {
                        "description": "Données des pays et de leurs positions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "countries": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Nom du pays"
                                                            },
                                                            "position": {
                                                                "type": "string",
                                                                "example": "Position géographique"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Une erreur interne s'est produite"
                    }
                }
            }
        },
        "/api/getStateInCountry": {
            "post": {
                "tags": [
                    "PaysEtatsVilles"
                ],
                "summary": "Obtenir les départements/États d'un pays",
                "operationId": "ad083a379ad2604dcbc70851c0489b2b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "country": {
                                        "type": "string",
                                        "example": "Nom du pays"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Liste des départements/États du pays",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "states": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Nom de l'État"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Une erreur interne s'est produite"
                    }
                }
            }
        },
        "/api/getCityInState": {
            "post": {
                "tags": [
                    "PaysEtatsVilles"
                ],
                "summary": "Obtenir les villes d'un département/État d'un pays",
                "operationId": "9204a5bafa922ca0f05af82d94e3a749",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "country": {
                                        "type": "string",
                                        "example": "Nom du pays"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "Nom du département/État"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Liste des villes du département/État",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "cities": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Nom de la ville"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Une erreur interne s'est produite"
                    }
                }
            }
        },
        "/api/getCityInCountry": {
            "post": {
                "tags": [
                    "PaysEtatsVilles"
                ],
                "summary": "Obtenir les villes d'un pays",
                "operationId": "b08c047bcfe04df07d3faa87f84e38b4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "country": {
                                        "type": "string",
                                        "example": "Nom du pays"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Liste des villes du pays",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "cities": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Nom de la ville"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Une erreur interne s'est produite"
                    }
                }
            }
        },
...
    },
...
}